Produce more readable Dot graphs
This removes the operator configuration from the output to make the
graphs we generate easier to read. It's not a perfect solution, however,
because going through petgraph::dot::Dot prohibits us from unwrapping
the `Option`s around nodes. I think we'll probably just want to
reimplement the dot printer from `petgraph` ourselves (it's not that
big), so that we can print exactly what we want.